Refactor logic deployment out
