+++ constants/deployments.json
@@ -1,15 +1,14 @@
 {
     "kovan": {
-        "ManualVolOracle": "0x93ad31E0f7C496B310Dc852f0e8a69E71C5a2864",
-        "OptionsPremiumPricerETH": "0x8C3c0b906A5b7bb896788BC8a19F9FC34eb4Ef38",
-        "OptionsPremiumPricerWBTC": "0x378C0152EdD6f79C74677F3E9251Eaee03B61861",
-        "RibbonThetaVaultETHCall": "0x35ABa03473793e878a6bb1B62426Bf9aCaa13C9C",
-        "RibbonThetaVaultETHCallLogic": "0xC7149F47D79314c02f5Ab852566854483342c1BC",
-        "RibbonThetaVaultETHPut": "0xD713F75652E687ff4522413d064c001598c45937",
-        "RibbonThetaVaultETHPutLogic": "0x5c92a8efd4aA4d75CfF0Db345E98c3AA8D50b0F3",
-        "RibbonThetaVaultWBTCCall": "0xe6f1559AcEE5063a5474ef319248Ab9E3fD54d7e",
-        "RibbonThetaVaultWBTCCallLogic": "0xdEab40DDB12E66FE85d5f0EE9B1B31Ea061eEb9A",
-        "StrikeSelectionETH": "0x5A40353F52497034C50C68c97579B3d9A406d994",
-        "VaultLifecycle": "0xDC2b9975D027C8b4C8C48F008582193692187bF3"
+        "ManualVolOracle": "0x6cD56377C7409693cf3D48e23B14EA72217bec22",
+        "OptionsPremiumPricerETH": "0x748F85AC6727EDc6b8120Fd7127a4Fb33d047d47",
+        "OptionsPremiumPricerWBTC": "0x36690d83a06971CD2F1f4A95052Ab0D5980e0FC8",
+        "RibbonThetaVaultETHCall": "0x6888cEaae27F1D66C4152082c186ed452E387FBB",
+        "RibbonThetaVaultETHPut": "0xc8B86C2dE51F124D99D783A11Be666E7c1Be2387",
+        "RibbonThetaVaultLogic": "0x739Ce06d2b435F81951fd5E4a5f67FD39E718b04",
+        "RibbonThetaVaultWBTCCall": "0xa6C65B92E24C72D5935396532c2536f0ca30654f",
+        "StrikeSelectionETH": "0xCf06D7a084739d869d3257F7190E6A75126F9e00",
+        "StrikeSelectionWBTC": "0x20D96F5754D39c75eA9094002DE2108e5F5d1329",
+        "VaultLifecycle": "0xe73ED716Af2C5fbB6b685009E11957f208507cf6"
     }
 }
