+++ deployments/kovan/ManualVolOracle.json
@@ -1,5 +1,5 @@
 {
-  "address": "0x93ad31E0f7C496B310Dc852f0e8a69E71C5a2864",
+  "address": "0x6cD56377C7409693cf3D48e23B14EA72217bec22",
   "abi": [
     {
       "inputs": [
@@ -329,22 +329,22 @@
       "type": "function"
     }
   ],
-  "transactionHash": "0x9f5abde2f2831a6162a4d529c90f63fa109f5fdfe12d3a5a9aafcf6bd6adf02a",
+  "transactionHash": "0xdb7bab71dd2b830e2941414820b315244d1ccd9e13db88daa6391fb25d93cb13",
   "receipt": {
     "to": null,
     "from": "0x8DD47c24aC72888BFb2b75c172bB55C127515884",
-    "contractAddress": "0x93ad31E0f7C496B310Dc852f0e8a69E71C5a2864",
-    "transactionIndex": 6,
+    "contractAddress": "0x6cD56377C7409693cf3D48e23B14EA72217bec22",
+    "transactionIndex": 5,
     "gasUsed": "634012",
-    "logsBloom": "0x00000004000000000000000001000000000000000000000000000000000000200020000000000000000000000100000000000000000000000000000000000000000100000000000000000000000000000000000000000000000000000000000000000000000000000000008000000000000010000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000100020000000000000000000009000000000000000000000000000000000000000000000000000000000000100000000000000000000000010000000000000000000000000000000000000000000000000000010",
-    "blockHash": "0x91ea973652452fddc266d5f13cebb335dbbcc2aed933f6939281ddd16bb83c95",
-    "transactionHash": "0x9f5abde2f2831a6162a4d529c90f63fa109f5fdfe12d3a5a9aafcf6bd6adf02a",
+    "logsBloom": "0x00000004000000000000000001000000000000000000000000000000000000200020000000000000000000000100000000000000000000000000000000000000000100000000000000000000000000800000000000000000000000000000000000000000000000000000008000000000000000000000000000000000000000000000000000000000000000000000000000000000000000800000000000000000000000000000000000000000100020000000000000000000009000000000000000000000000000000000000000000000000000000000000100000000080000000000000000000000000000000000000000000000000000000000000000000000",
+    "blockHash": "0x0e62db8f8d749c3166c6f0f366c1565431e9de39772e5eef3ca817651c6b7dac",
+    "transactionHash": "0xdb7bab71dd2b830e2941414820b315244d1ccd9e13db88daa6391fb25d93cb13",
     "logs": [
       {
-        "transactionIndex": 6,
-        "blockNumber": 26711128,
-        "transactionHash": "0x9f5abde2f2831a6162a4d529c90f63fa109f5fdfe12d3a5a9aafcf6bd6adf02a",
-        "address": "0x93ad31E0f7C496B310Dc852f0e8a69E71C5a2864",
+        "transactionIndex": 5,
+        "blockNumber": 26746378,
+        "transactionHash": "0xdb7bab71dd2b830e2941414820b315244d1ccd9e13db88daa6391fb25d93cb13",
+        "address": "0x6cD56377C7409693cf3D48e23B14EA72217bec22",
         "topics": [
           "0x2f8788117e7eff1d82e926ec794901d17c78024a50270940304540a733656f0d",
           "0xdf8b4c520ffe197c5343c6f5aec59570151ef9a492f2c624fd45ddde6135ec42",
@@ -352,12 +352,12 @@
           "0x0000000000000000000000008dd47c24ac72888bfb2b75c172bb55c127515884"
         ],
         "data": "0x",
-        "logIndex": 8,
-        "blockHash": "0x91ea973652452fddc266d5f13cebb335dbbcc2aed933f6939281ddd16bb83c95"
+        "logIndex": 11,
+        "blockHash": "0x0e62db8f8d749c3166c6f0f366c1565431e9de39772e5eef3ca817651c6b7dac"
       }
     ],
-    "blockNumber": 26711128,
-    "cumulativeGasUsed": "5228702",
+    "blockNumber": 26746378,
+    "cumulativeGasUsed": "1074067",
     "status": 1,
     "byzantium": true
   },
