+++ deployments/kovan/OptionsPremiumPricerETH.json
@@ -1,5 +1,5 @@
 {
-  "address": "0x8C3c0b906A5b7bb896788BC8a19F9FC34eb4Ef38",
+  "address": "0x748F85AC6727EDc6b8120Fd7127a4Fb33d047d47",
   "abi": [
     {
       "inputs": [
@@ -180,25 +180,25 @@
       "type": "function"
     }
   ],
-  "transactionHash": "0x2ee479af5c0a3413ccba4b2e32987a4d8f1676c428ab7080080eaccd93d2c02c",
+  "transactionHash": "0xe226b69d2ad1966a61f1b39bed801718ac2e478a1482d2fe711f796e204f7669",
   "receipt": {
     "to": null,
     "from": "0x8DD47c24aC72888BFb2b75c172bB55C127515884",
-    "contractAddress": "0x8C3c0b906A5b7bb896788BC8a19F9FC34eb4Ef38",
+    "contractAddress": "0x748F85AC6727EDc6b8120Fd7127a4Fb33d047d47",
     "transactionIndex": 0,
     "gasUsed": "1186540",
     "logsBloom": "0x00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000",
-    "blockHash": "0x1e9a58c9699dec5fc4dbb7aa839e0308abf06afd399f0daee6c7bd698112faf6",
-    "transactionHash": "0x2ee479af5c0a3413ccba4b2e32987a4d8f1676c428ab7080080eaccd93d2c02c",
+    "blockHash": "0x44a6bc29380c02d4c4e7f5e498cb41aa65397b703ba52474f4fbe6585321d0f8",
+    "transactionHash": "0xe226b69d2ad1966a61f1b39bed801718ac2e478a1482d2fe711f796e204f7669",
     "logs": [],
-    "blockNumber": 26711152,
+    "blockNumber": 26746394,
     "cumulativeGasUsed": "1186540",
     "status": 1,
     "byzantium": true
   },
   "args": [
     "0x8ad599c3A0ff1De082011EFDDc58f1908eb6e6D8",
-    "0x93ad31E0f7C496B310Dc852f0e8a69E71C5a2864",
+    "0x6cD56377C7409693cf3D48e23B14EA72217bec22",
     "0x0c15Ab9A0DB086e062194c273CC79f41597Bbf13",
     "0x9211c6b3BF41A10F78539810Cf5c64e1BB78Ec60"
   ],
