+++ deployments/kovan/OptionsPremiumPricerWBTC.json
@@ -1,5 +1,5 @@
 {
-  "address": "0x378C0152EdD6f79C74677F3E9251Eaee03B61861",
+  "address": "0x36690d83a06971CD2F1f4A95052Ab0D5980e0FC8",
   "abi": [
     {
       "inputs": [
@@ -180,25 +180,25 @@
       "type": "function"
     }
   ],
-  "transactionHash": "0xb4ec044bbf630135563c1adf7db2afc4a001f2db65902bd74516104bcebfffa4",
+  "transactionHash": "0x5eda1598097396dbf91fc6f12cc35c1236abc23b25b8b2007f1215769a1fb7f0",
   "receipt": {
     "to": null,
     "from": "0x8DD47c24aC72888BFb2b75c172bB55C127515884",
-    "contractAddress": "0x378C0152EdD6f79C74677F3E9251Eaee03B61861",
-    "transactionIndex": 4,
+    "contractAddress": "0x36690d83a06971CD2F1f4A95052Ab0D5980e0FC8",
+    "transactionIndex": 0,
     "gasUsed": "1186540",
     "logsBloom": "0x00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000",
-    "blockHash": "0xc90996a961df69a2f1298efec59d3330e19226fe57ea65d00d612584c1224bff",
-    "transactionHash": "0xb4ec044bbf630135563c1adf7db2afc4a001f2db65902bd74516104bcebfffa4",
+    "blockHash": "0x299d8692c888fc8c97dfb2649e3a054b3485769417c0887d4898b3c3c8990c26",
+    "transactionHash": "0x5eda1598097396dbf91fc6f12cc35c1236abc23b25b8b2007f1215769a1fb7f0",
     "logs": [],
-    "blockNumber": 26711142,
-    "cumulativeGasUsed": "5767178",
+    "blockNumber": 26746388,
+    "cumulativeGasUsed": "1186540",
     "status": 1,
     "byzantium": true
   },
   "args": [
     "0x99ac8cA7087fA4A2A1FB6357269965A2014ABc35",
-    "0x93ad31E0f7C496B310Dc852f0e8a69E71C5a2864",
+    "0x6cD56377C7409693cf3D48e23B14EA72217bec22",
     "0x6135b13325bfC4B00278B4abC5e20bbce2D6580e",
     "0x9211c6b3BF41A10F78539810Cf5c64e1BB78Ec60"
   ],
