+++ deployments/kovan/VaultLifecycle.json
@@ -1,5 +1,5 @@
 {
-  "address": "0xDC2b9975D027C8b4C8C48F008582193692187bF3",
+  "address": "0xe73ED716Af2C5fbB6b685009E11957f208507cf6",
   "abi": [
     {
       "inputs": [
@@ -121,19 +121,19 @@
       "type": "function"
     }
   ],
-  "transactionHash": "0x4d0fa3237c98e0a6780c8d15e134b9fe975f4239f86b7e3b06272077e9ebee40",
+  "transactionHash": "0x338c53b946a4ba99e8ec099f1e2b929726a8f07d31667c0da2c24d83410cf8f9",
   "receipt": {
     "to": null,
     "from": "0x8DD47c24aC72888BFb2b75c172bB55C127515884",
-    "contractAddress": "0xDC2b9975D027C8b4C8C48F008582193692187bF3",
+    "contractAddress": "0xe73ED716Af2C5fbB6b685009E11957f208507cf6",
     "transactionIndex": 2,
     "gasUsed": "3274087",
     "logsBloom": "0x00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000",
-    "blockHash": "0xbb8f166c2d58b6c2f4c10668ee01ef8c787f85d21b06ef0c07691869adab9c40",
-    "transactionHash": "0x4d0fa3237c98e0a6780c8d15e134b9fe975f4239f86b7e3b06272077e9ebee40",
+    "blockHash": "0xd4596cbf8f24621edaf706a7de6d426ae3e63b13d9ec5412a1fc20fe5fbacec3",
+    "transactionHash": "0x338c53b946a4ba99e8ec099f1e2b929726a8f07d31667c0da2c24d83410cf8f9",
     "logs": [],
-    "blockNumber": 26711135,
-    "cumulativeGasUsed": "7163241",
+    "blockNumber": 26746380,
+    "cumulativeGasUsed": "3557665",
     "status": 1,
     "byzantium": true
   },
