+++ scripts/deploy/04_eth_put_theta_vault.ts
@@ -5,20 +5,12 @@ import {
   MAINNET_USDC_ORACLE,
   KOVAN_USDC_ORACLE,
   USDC_ADDRESS,
-  OTOKEN_FACTORY,
-  OTOKEN_FACTORY_KOVAN,
-  GAMMA_CONTROLLER,
-  GAMMA_CONTROLLER_KOVAN,
-  MARGIN_POOL,
-  MARGIN_POOL_KOVAN,
-  GNOSIS_EASY_AUCTION,
-  EASY_AUCTION_KOVAN,
 } from "../../constants/constants";
 import OptionsPremiumPricer_ABI from "../../constants/abis/OptionsPremiumPricer.json";
 
 const ETH_USDC_POOL = "0x8ad599c3A0ff1De082011EFDDc58f1908eb6e6D8";
 const MAINNET_ETH_ORACLE = "0x5f4eC3Df9cbd43714FE2740f5E3616155c5b8419";
-const KOVAN_ETH_ORACLE = "0x9326BFA02ADD2366b30bacB125260Af641031331";
+const KOVAN_ETH_ORACLE = "0x0c15Ab9A0DB086e062194c273CC79f41597Bbf13";
 const KOVAN_WETH = "0xd0A1E359811322d97991E03f863a0C30C2cF029C";
 const KOVAN_USDC = "0x7e6edA50d1c833bE936492BF42C1BF376239E9e2";
 
@@ -35,7 +27,7 @@ const main = async ({
   const { parseEther } = ethers.utils;
   const { deploy } = deployments;
   const { deployer, owner, admin, feeRecipient } = await getNamedAccounts();
-  console.log(`01 - Deploying ETH Call Theta Vault on ${network.name}`);
+  console.log(`04 - Deploying ETH Put Theta Vault on ${network.name}`);
 
   const isMainnet = network.name === "mainnet";
   const manualVolOracle = await deployments.get("ManualVolOracle");
@@ -63,28 +55,11 @@ const main = async ({
     args: [pricer.address, STRIKE_DELTA, STRIKE_STEP],
   });
 
-  const lifecycle = await deploy("VaultLifecycle", {
-    contract: "VaultLifecycle",
-    from: deployer,
-  });
-
   const weth = isMainnet ? WETH_ADDRESS : KOVAN_WETH;
+  const usdc = isMainnet ? USDC_ADDRESS : KOVAN_USDC;
 
-  const logicDeployment = await deploy("RibbonThetaVaultETHCallLogic", {
-    contract: "RibbonThetaVault",
-    from: deployer,
-    args: [
-      weth,
-      isMainnet ? USDC_ADDRESS : KOVAN_USDC,
-      isMainnet ? OTOKEN_FACTORY : OTOKEN_FACTORY_KOVAN,
-      isMainnet ? GAMMA_CONTROLLER : GAMMA_CONTROLLER_KOVAN,
-      isMainnet ? MARGIN_POOL : MARGIN_POOL_KOVAN,
-      isMainnet ? GNOSIS_EASY_AUCTION : EASY_AUCTION_KOVAN,
-    ],
-    libraries: {
-      VaultLifecycle: lifecycle.address,
-    },
-  });
+  const logicDeployment = await deployments.get("RibbonThetaVaultLogic");
+  const lifecycle = await deployments.get("VaultLifecycle");
 
   const RibbonThetaVault = await ethers.getContractFactory("RibbonThetaVault", {
     libraries: {
@@ -97,19 +72,19 @@ const main = async ({
     feeRecipient,
     0,
     0,
-    "Ribbon ETH Theta Vault",
-    "rETH-THETA",
+    "Ribbon USDC Theta Vault ETH Put",
+    "rUSDC-ETH-P-THETA",
     pricer.address,
     strikeSelection.address,
-    50, //5% discount
+    50, // 5% discount
     3600, // 1 hour auction duration
     {
       isPut: false,
-      decimals: 18,
-      asset: weth,
+      decimals: 6,
+      asset: usdc,
       underlying: weth,
-      minimumSupply: BigNumber.from(10).pow(10),
-      cap: parseEther("1000"),
+      minimumSupply: BigNumber.from(10).pow(3),
+      cap: parseEther("1000000"),
     },
   ];
   const initData = RibbonThetaVault.interface.encodeFunctionData(
@@ -117,13 +92,13 @@ const main = async ({
     initArgs
   );
 
-  await deploy("RibbonThetaVaultETHCall", {
+  await deploy("RibbonThetaVaultETHPut", {
     contract: "AdminUpgradeabilityProxy",
     from: deployer,
     args: [logicDeployment.address, admin, initData],
   });
 };
-main.tags = ["RibbonThetaVaultETHCall"];
-main.dependencies = ["ManualVolOracle"];
+main.tags = ["RibbonThetaVaultETHPut"];
+main.dependencies = ["ManualVolOracle", "RibbonThetaVaultLogic"];
 
 export default main;
